Consolidate styling into a single easy-to-import file
