--- packages/itinerary-body/src/RouteBadge/styled-accessibility.js
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const SROnly = styled.span`
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
-`;
-
-export const SRHidden = styled.span.attrs({ "aria-hidden": true })``;
